renames action and reducer files to use camelcase
